+++ Ivna Ivankovic/SA-DS/bucket.h
@@ -7,10 +7,10 @@
 typedef struct
 {
     char *substring;
-    size_t length;
+    int length;
 } bucket_t;
 
-bucket_t *create_bucket(char *substring, const size_t length);
+bucket_t *create_bucket(char *substring, const int length);
 
 int bucket_compare(bucket_t *first, bucket_t *second);
