+++ Ivna Ivankovic/SA-DS/bucket_array.h
@@ -7,16 +7,16 @@
 typedef struct
 {
     bucket_t *buckets;
-    size_t size;
+    int size;
 } bucket_array_t;
 
-bucket_array_t *create_bucket_array(const size_t size);
+bucket_array_t *create_bucket_array(const int size);
 
 void free_bucket_array(bucket_array_t *bucket_array);
 
-bucket_t *bucket_at(bucket_array_t *bucket_array, const size_t index);
+bucket_t *bucket_at(bucket_array_t *bucket_array, const int index);
 
-int adjust_bucket_array_size(bucket_array_t *bucket_array, const size_t new_size);
+int adjust_bucket_array_size(bucket_array_t *bucket_array, const int new_size);
 
 void sort_bucket_array(bucket_array_t *bucket_array);
